ISAICP-4518: Rename '3_way_merge' to 'import' and 'entity_fields_values' to '3_way_merge'.
