+++ web/modules/custom/joinup_federation/src/Plugin/pipeline/Step/ThreeWayMerge.php
@@ -7,20 +7,20 @@ namespace Drupal\joinup_federation\Plugin\pipeline\Step;
 use Drupal\Core\Entity\EntityFieldManagerInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\joinup_federation\JoinupFederationStepPluginBase;
-use Drupal\pipeline\Plugin\PipelineStepWithBatchInterface;
 use Drupal\pipeline\Plugin\PipelineStepWithBatchTrait;
+use Drupal\pipeline\Plugin\PipelineStepWithBatchInterface;
 use Drupal\rdf_entity\Database\Driver\sparql\Connection;
 use Drupal\rdf_entity\Entity\Rdf;
 use Drupal\rdf_entity\RdfEntityGraphInterface;
-use Drupal\rdf_schema_field_validation\SchemaFieldValidatorInterface;
+use Drupal\rdf_entity\RdfInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Merges the incoming data with local data.
+ * Ensures values for empty fields from the staging graph.
  *
  * @PipelineStep(
  *   id = "3_way_merge",
- *   label = @Translation("3-way merge"),
+ *   label = @Translation("Assign default or local values to empty fields of imported entities"),
  * )
  */
 class ThreeWayMerge extends JoinupFederationStepPluginBase implements PipelineStepWithBatchInterface {
@@ -43,13 +43,6 @@ class ThreeWayMerge extends JoinupFederationStepPluginBase implements PipelineSt
   protected $entityFieldManager;
 
   /**
-   * The RDF schema field validator service.
-   *
-   * @var \Drupal\rdf_schema_field_validation\SchemaFieldValidatorInterface
-   */
-  protected $rdfSchemaFieldValidator;
-
-  /**
    * Creates a new pipeline step plugin instance.
    *
    * @param array $configuration
@@ -64,14 +57,11 @@ class ThreeWayMerge extends JoinupFederationStepPluginBase implements PipelineSt
    *   The entity type manager service.
    * @param \Drupal\Core\Entity\EntityFieldManagerInterface $entity_field_manager
    *   The entity field manager service.
-   * @param \Drupal\rdf_schema_field_validation\SchemaFieldValidatorInterface $rdf_schema_field_validator
-   *   The RDF schema field validator service.
    */
-  public function __construct(array $configuration, $plugin_id, array $plugin_definition, Connection $sparql, EntityTypeManagerInterface $entity_type_manager, EntityFieldManagerInterface $entity_field_manager, SchemaFieldValidatorInterface $rdf_schema_field_validator) {
+  public function __construct(array $configuration, $plugin_id, array $plugin_definition, Connection $sparql, EntityTypeManagerInterface $entity_type_manager, EntityFieldManagerInterface $entity_field_manager) {
     parent::__construct($configuration, $plugin_id, $plugin_definition, $sparql);
     $this->entityTypeManager = $entity_type_manager;
     $this->entityFieldManager = $entity_field_manager;
-    $this->rdfSchemaFieldValidator = $rdf_schema_field_validator;
   }
 
   /**
@@ -84,97 +74,231 @@ class ThreeWayMerge extends JoinupFederationStepPluginBase implements PipelineSt
       $plugin_definition,
       $container->get('sparql_endpoint'),
       $container->get('entity_type.manager'),
-      $container->get('entity_field.manager'),
-      $container->get('rdf_schema_field_validation.schema_field_validator')
+      $container->get('entity_field.manager')
     );
   }
 
   /**
    * {@inheritdoc}
    */
+  public function defaultConfiguration() {
+    return ['collection' => NULL] + parent::defaultConfiguration();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
   public function initBatchProcess() {
-    // Retrieve the list of entities from the persistent data store as an
-    // associative array keyed by entity ID and having a boolean as value,
-    // signaling if the entity already exists in Joinup.
-    $remaining_entity_ids = $this->getPersistentDataValue('entities');
-    $this->setBatchValue('remaining_entity_ids', $remaining_entity_ids);
-    // Estimate the number of iterations.
-    return ceil(count($remaining_entity_ids) / static::BATCH_SIZE);
+    $whitelist = $this->getPersistentDataValue('whitelist');
+    $this->unsetPersistentDataValue('whitelist');
+
+    // Get the incoming entities that are stored also locally.
+    $local_ids = $this->getSparqlQuery()
+      // @todo Remove call to ::graphs() in ISAICP-4497.
+      // @see https://webgate.ec.europa.eu/CITnet/jira/browse/ISAICP-4497
+      ->graphs(['default', 'draft'])
+      ->condition('id', array_values($whitelist), 'IN')
+      ->execute();
+
+    $incoming_ids = [];
+    foreach ($whitelist as $id) {
+      $incoming_ids[$id] = isset($local_ids[$id]);
+    }
+
+    $this->setBatchValue('remaining_incoming_ids', $incoming_ids);
+
+    // Estimating the total number of iterations.
+    return ceil(count($incoming_ids) / static::BATCH_SIZE);
   }
 
   /**
    * {@inheritdoc}
    */
   public function batchProcessIsCompleted() {
-    return !$this->getBatchValue('remaining_entity_ids');
+    return !$this->getBatchValue('remaining_incoming_ids');
   }
 
   /**
    * {@inheritdoc}
    */
   public function execute() {
-    $ids_to_process = $this->extractNextSubset('remaining_entity_ids', static::BATCH_SIZE);
-    $ids = array_keys($ids_to_process);
+    $ids_to_process = $this->extractNextSubset('remaining_incoming_ids', static::BATCH_SIZE);
+    $incoming_ids = array_keys($ids_to_process);
+    $local_ids = array_keys(array_filter($ids_to_process));
 
-    // Build a list of local entities that are about to be updated.
-    $local_entity_ids = array_keys(array_filter($ids_to_process));
-    /** @var \Drupal\rdf_entity\RdfInterface[] $local_entities */
+    /** @var \Drupal\rdf_entity\RdfInterface[] $incoming_entities */
+    $incoming_entities = $incoming_ids ? $this->getRdfStorage()->loadMultiple($incoming_ids, ['staging']) : [];
     // @todo Remove the 2nd argument of ::loadMultiple() in ISAICP-4497.
     // @see https://webgate.ec.europa.eu/CITnet/jira/browse/ISAICP-4497
-    $local_entities = $local_entity_ids ? Rdf::loadMultiple($local_entity_ids, [RdfEntityGraphInterface::DEFAULT, 'draft']) : [];
+    $local_entities = $local_ids ? Rdf::loadMultiple($local_ids, [RdfEntityGraphInterface::DEFAULT, 'draft']) : [];
 
-    $entities_to_save = $entities_to_delete = [];
-    /** @var \Drupal\rdf_entity\RdfInterface $entity */
-    foreach (Rdf::loadMultiple($ids, ['staging']) as $id => $entity) {
-      // This entity already exists.
-      if ($ids_to_process[$id]) {
-        $graph_ids = [];
-        foreach ([RdfEntityGraphInterface::DEFAULT, 'draft'] as $graph_id) {
-          if ($local_entities[$id]->hasGraph($graph_id)) {
-            $graph_ids[$graph_id] = $graph_id;
-          }
+    // Collect here entity IDs that are about to be saved.
+    $entities = $this->hasPersistentDataValue('entities') ? $this->getPersistentDataValue('entities') : [];
+
+    /** @var \Drupal\rdf_entity\RdfInterface $incoming_entity */
+    foreach ($incoming_entities as $id => $incoming_entity) {
+      $entity_exists = isset($local_entities[$id]);
+      // The entity already exists.
+      if ($entity_exists) {
+        // Check for bundle mismatch between the incoming and the local entity.
+        if ($incoming_entity->bundle() !== $local_entities[$id]->bundle()) {
+          $arguments = [
+            '%id' => $id,
+            '@incoming' => $incoming_entity->get('rid')->entity->getSingularLabel(),
+            '@local' => $local_entities[$id]->get('rid')->entity->getSingularLabel(),
+          ];
+          return [
+            '#markup' => $this->t("The imported @incoming with the ID '%id' tries to override a local @local with the same ID.", $arguments),
+          ];
         }
 
-        // Pick up first graph to be set.
-        $graph_id = key($graph_ids);
-        $local_entity = clone $entity;
-        $local_entity->set('graph', $graph_id);
-
-        // If the local entity exists in both, 'default' and 'draft', graphs, we
-        // remove the 'draft' version. This is needed because the federated
-        // fields cannot be edited locally anymore and a potential publish of
-        // the draft could override the federated fields values.
-        if (count($graph_ids) > 1) {
-          $this->getRdfStorage()->deleteFromGraph([$local_entity], 'draft');
-        }
-
-        // Collect the entities to be deleted later from the 'staging' graph. We
-        // are not deleting here because, when saving the entities in the main
-        // graphs, this would lead to a null $entity->original.
-        $entities_to_delete[] = $entity;
+        $needs_save = $this->updateStagingFieldsFromLocalValues($incoming_entity, $local_entities[$id]);
       }
-      // No local entity. Copy the incoming entity as a published entity.
+      // No local entity. Ensure defaults.
       else {
-        $local_entity = (clone $entity)
-          ->enforceIsNew()
-          ->set('graph', RdfEntityGraphInterface::DEFAULT);
-        // Delete the incoming entity from the staging graph.
-        $entity->skip_notification = TRUE;
-        $entity->delete();
+        $this->ensureFieldDefaults($incoming_entity);
+        $needs_save = TRUE;
       }
-      $entities_to_save[] = $local_entity;
+
+      if ($needs_save) {
+        $this->handleAffiliation($incoming_entity, $entity_exists);
+        $incoming_entity->skip_notification = TRUE;
+        $incoming_entity->save();
+        $entities[$incoming_entity->id()] = $entity_exists;
+      }
     }
 
-    // Save the entities.
-    foreach ($entities_to_save as $local_entity) {
-      $local_entity->skip_notification = TRUE;
-      $local_entity->save();
+    // Persist the list so we can reuse it in the next steps.
+    $this->setPersistentDataValue('entities', $entities);
+  }
+
+  /**
+   * Updates the field values from the local to the incoming entity.
+   *
+   * @param \Drupal\rdf_entity\RdfInterface $incoming_entity
+   *   The imported entity.
+   * @param \Drupal\rdf_entity\RdfInterface $local_entity
+   *   The local entity.
+   *
+   * @return bool
+   *   If the incoming entity has been changed and needs to be saved.
+   */
+  protected function updateStagingFieldsFromLocalValues(RdfInterface $incoming_entity, RdfInterface $local_entity): bool {
+    $changed = FALSE;
+
+    foreach ($incoming_entity->getFieldDefinitions() as $field_name => $field_definition) {
+      // Bypass fields without mapping or fields we don't want to override.
+      if (in_array($field_name, ['id', 'rid', 'graph', 'uuid', 'uid'])) {
+        continue;
+      }
+      // Only stored fields are allowed.
+      if ($field_definition->isComputed()) {
+        continue;
+      }
+
+      $columns = $field_definition->getFieldStorageDefinition()->getColumns();
+      foreach ($columns as $column_name => $column_schema) {
+        // Check if the field is an ADMS-AP field.
+        $incoming_field = $incoming_entity->get($field_name);
+        $local_field = $local_entity->get($field_name);
+        // Assign only if the incoming field is empty.
+        if ($incoming_field->isEmpty()) {
+          $incoming_field->setValue($local_field->getValue());
+          $changed = TRUE;
+          // Don't check the rest of the columns because the whole field has
+          // been already assigned.
+          break;
+        }
+      }
     }
 
-    // Cleanup the entities from the 'staging' graph.
-    if ($entities_to_delete) {
-      $this->getRdfStorage()->deleteFromGraph($entities_to_delete, 'staging');
+    return $changed;
+  }
+
+  /**
+   * Sets default values for fields.
+   *
+   * @param \Drupal\rdf_entity\RdfInterface $incoming_entity
+   *   The imported entity.
+   */
+  protected function ensureFieldDefaults(RdfInterface $incoming_entity): void {
+    // Determine the state field for this bundle, if any.
+    $state_field_name = NULL;
+    $state_field_map = $this->entityFieldManager->getFieldMapByFieldType('state');
+    if (!empty($state_field_map['rdf_entity'])) {
+      foreach ($state_field_map['rdf_entity'] as $field_name => $field_info) {
+        if (isset($field_info['bundles'][$incoming_entity->bundle()])) {
+          $state_field_name = $field_name;
+          break;
+        }
+      }
     }
+
+    // Set he state of new entities to 'validated'.
+    if ($state_field_name) {
+      $incoming_entity->set($state_field_name, 'validated');
+    }
+
+    /** @var \Drupal\Core\Field\FieldItemListInterface $field */
+    foreach ($incoming_entity as $field_name => &$field) {
+      // Populate empty fields with their default value. This is a Drupal
+      // content entity that was not created via Drupal API. As an effect, empty
+      // fields didn't receive their default values. We have to explicitly do
+      // this before saving. The check if the field is a computed field is there
+      // in order to avoid the computation of the field. The method ::isEmpty()
+      // triggers the computation.
+      // @see \Drupal\Core\TypedData\ComputedItemListTrait::isEmpty()
+      if (!$field->getFieldDefinition()->isComputed() && $field->isEmpty()) {
+        $field->applyDefaultValue();
+      }
+    }
+  }
+
+  /**
+   * Handles the incoming solution affiliation.
+   *
+   * For existing solutions, we only check if the configured collection ID
+   * matches the solution affiliation. For new solutions, we affiliate the
+   * solution to the configured collection.
+   *
+   * @param \Drupal\rdf_entity\RdfInterface $incoming_solution
+   *   The local solution.
+   * @param bool $entity_exists
+   *   If the incoming entity already exits on the system.
+   *
+   * @throws \Exception
+   *   If the configured collection is different than the collection of the
+   *   local solution.
+   */
+  protected function handleAffiliation(RdfInterface $incoming_solution, bool $entity_exists): void {
+    // Check only solutions.
+    if ($incoming_solution->bundle() !== 'solution') {
+      return;
+    }
+
+    // If this plugin was not configured to assign a collection, exit early.
+    if (!$collection_id = $this->getConfiguration()['collection']) {
+      return;
+    }
+
+    if (!$entity_exists) {
+      $incoming_solution->set('collection', $collection_id);
+      return;
+    }
+
+    // Check for collection mismatch when federating an existing solution.
+    $match = FALSE;
+    foreach ($incoming_solution->get('collection') as $item) {
+      if ($item->target_id === $collection_id) {
+        $match = TRUE;
+        break;
+      }
+    }
+
+    if (!$match) {
+      throw new \Exception("Plugin '3_way_merge' is configured to assign the '$collection_id' collection but the existing solution '{$incoming_solution->id()}' has '{$incoming_solution->collection->target_id}' as collection.");
+    }
+    // For an existing solution we don't make any changes to its affiliation.
   }
 
 }
